refactor: split models/bayesian_ets.py (854L) into subpackage (#224)
- bayesian_ets/__init__.py (33L) — re-exports all 12 public + tested names
- bayesian_ets/priors.py (52L) — ETSPriors, _log_prior_* functions
- bayesian_ets/inference.py (364L) — loglik, pack/unpack, log_posterior, MAP, MCMC, forecast
- bayesian_ets/model.py (223L) — BayesianETSResult, BayesianETS, bayesian_ets()
- All files under 500 lines (max 364L vs original 854L)
- All 12 existing import paths preserved via __init__.py re-exports
- 6 new structural tests, 65/65 total tests pass
